acacetatest5/backend/requirements.txt
Flask==2.2.2
Flask_SQLAlchemy==2.5.1
Flask_Bcrypt==0.7.1
Flask_JWT_Extended==4.4.4
Flask_Migrate==3.1.0
Flask_Cors==3.0.10
opensearch-py==2.0.0
sentence-transformers==2.2.2
pinecone-client==2.1.1
pytest==7.1.2
pytest-flask==1.2.0
mysql-connector-python==8.0.29
